requirements.txt
terminaltables
six
addict
numba  
tqdm
pyquaternion          
nuscenes-devkit
einops    
nvtx      
onnx
pandas
torchvision==0.15.1
--find-links https://data.pyg.org/whl/torch-2.1.0+cu118.html
torch-scatter
# tensorrt==8.5.1.7
